IoT hardware emulator script
This commit introduces a script to emulate the hardware
sensors data. It generates random values for the 2 sensors,
and only generates "concerning" values for a percentage
of the time.

The script sends this data to a user entered IP:PORT

Usage
-----
Linux: ./harware_emulator.py -v -i <IP> -p <PORT>