Introduce HW unit testing
This commit introduces unit tests for the hardware and
hardware emulator scripts. This commit introduces some
structure changes to these scripts to make them more
modular and easier to test.

Instructions for testing are in the README file in the
test directory. An environment file needs to be
sourced before running tests.