feat: separate handling for LED
This commit introduces two LED icons to be
displayed on the SenseHAT LED screen (red
for warning, green for safe).

This commit also introduces the basic messaging
scheme and underlying handling for the
server node to instruct the Pi.

This commit also introduces the changes to
the emulator script. However, this commit
does not account for determining the IP
address to send the message back to the Pi.

This change shouldn't break any current functionality.